sobel-service/requirements.txt
opencv-python
flask
google-cloud-storage
pika